requirements.txt
pygame-ce
pymunk
google-generativeai
watchdog
pygame_gui
